requirements.txt
fastapi
uvicorn
pydantic
requests
supabase
python-dotenv
jinja2
google-generativeai